# contract/openapi.json
{
  "components": {
...
    "schemas": {
...
      "Application": {
        "properties": {
          "answersId": {
            "type": [
              "string",
              "null"
            ]
          },
          "appliedAt": {
            "format": "date-time",
            "type": "string"
          },
          "company": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "jobId": {
            "type": "string"
          },
          "meta": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "score": {
            "maximum": 5,
            "minimum": 0,
            "type": "number"
          },
          "stage": {
            "anyOf": [
              {
                "enum": [
                  0
                ],
                "type": "number"
              },
              {
                "enum": [
                  1
                ],
                "type": "number"
              },
              {
                "enum": [
                  2
                ],
                "type": "number"
              },
              {
                "enum": [
                  3
                ],
                "type": "number"
              }
            ]
          },
          "statusLabel": {
            "type": "string"
          },
          "statusTone": {
            "enum": [
              "good",
              "verified",
              "neutral"
            ],
            "type": "string"
          },
          "tailored": {
            "type": "boolean"
          },
          "tailoredResumeId": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "jobId",
          "role",
          "company",
          "meta",
          "appliedAt",
          "score",
          "stage",
          "statusLabel",
          "statusTone",
          "tailored",
          "note",
          "tailoredResumeId",
          "answersId"
        ],
        "type": "object"
      },
...
      "Eligibility": {
        "properties": {
          "summary": {
            "type": "string"
          },
          "tier": {
            "$ref": "#/components/schemas/EligibilityTier"
          },
          "tone": {
            "$ref": "#/components/schemas/Tone"
          }
        },
        "required": [
          "tier",
          "tone",
          "summary"
        ],
        "type": "object"
      },
      "EligibilityTier": {
        "enum": [
          "anywhere",
          "eligible",
          "local",
          "abroad",
          "unknown"
        ],
        "type": "string"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "enum": [
                  "VALIDATION_ERROR",
                  "NOT_FOUND",
                  "CONFLICT",
                  "RUN_NOT_READY",
                  "PARSE_FAILED",
                  "EXTRACTION_FAILED",
                  "UPSTREAM_LLM_ERROR",
                  "PAYLOAD_TOO_LARGE",
                  "FETCH_BLOCKED",
                  "NOT_A_JOB_POSTING",
                  "INTERNAL",
                  "UNAUTHORIZED",
                  "FORBIDDEN"
                ],
                "type": "string"
              },
              "details": {},
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
      "Job": {
        "properties": {
          "applyUrl": {
            "format": "uri",
            "type": "string"
          },
          "breakdown": {
            "items": {
              "properties": {
                "display": {
                  "type": "string"
                },
                "label": {
                  "type": "string"
                },
                "tone": {
                  "$ref": "#/components/schemas/Tone"
                },
                "value": {
                  "type": "number"
                }
              },
              "required": [
                "label",
                "value"
              ],
              "type": "object"
            },
            "type": "array"
          },
          "company": {
            "type": "string"
          },
          "eligibility": {
            "$ref": "#/components/schemas/Eligibility"
          },
          "firstSeen": {
            "format": "date-time",
            "type": "string"
          },
          "fit": {
            "items": {
              "properties": {
                "k": {
                  "type": "string"
                },
                "v": {
                  "type": "string"
                }
              },
              "required": [
                "k",
                "v"
              ],
              "type": "object"
            },
            "type": "array"
          },
          "gaps": {
            "items": {
              "properties": {
                "k": {
                  "type": "string"
                },
                "tone": {
                  "enum": [
                    "warn",
                    "ok"
                  ],
                  "type": "string"
                },
                "v": {
                  "type": "string"
                }
              },
              "required": [
                "tone",
                "k",
                "v"
              ],
              "type": "object"
            },
            "type": "array"
          },
          "ghost": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "isNew": {
            "type": "boolean"
          },
          "legitimacy": {
            "$ref": "#/components/schemas/Legitimacy"
          },
          "meta": {
            "type": "string"
          },
          "persona": {
            "$ref": "#/components/schemas/Persona"
          },
          "role": {
            "type": "string"
          },
          "score": {
            "maximum": 5,
            "minimum": 0,
            "type": "number"
          },
          "source": {
            "$ref": "#/components/schemas/SourceRef"
          },
          "tags": {
            "items": {
              "properties": {
                "label": {
                  "type": "string"
                },
                "title": {
                  "type": "string"
                },
                "tone": {
                  "$ref": "#/components/schemas/Tone"
                }
              },
              "required": [
                "tone",
                "label"
              ],
              "type": "object"
            },
            "type": "array"
          },
          "verdict": {
            "type": "string"
          },
          "why": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "score",
          "role",
          "company",
          "meta",
          "verdict",
          "why",
          "tags",
          "breakdown",
          "fit",
          "gaps",
          "legitimacy",
          "eligibility",
          "applyUrl",
          "source",
          "persona",
          "firstSeen",
          "isNew"
        ],
        "type": "object"
      },
      "Legitimacy": {
        "properties": {
          "confidence": {
            "maximum": 1,
            "minimum": 0,
            "type": "number"
          },
          "summary": {
            "type": "string"
          },
          "tier": {
            "$ref": "#/components/schemas/LegitimacyTier"
          },
          "tone": {
            "$ref": "#/components/schemas/Tone"
          },
          "webEvidence": {
            "oneOf": [
              {
                "properties": {
                  "companySignals": {
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "confidence": {
                    "maximum": 1,
                    "minimum": 0,
                    "type": "number"
                  },
                  "sightings": {
                    "items": {
                      "properties": {
                        "postedDate": {
                          "type": "string"
                        },
                        "source": {
                          "type": "string"
                        },
                        "url": {
                          "format": "uri",
                          "type": "string"
                        }
                      },
                      "required": [
                        "url",
                        "source"
                      ],
                      "type": "object"
                    },
                    "type": "array"
                  },
                  "status": {
                    "enum": [
                      "ok"
                    ],
                    "type": "string"
                  },
                  "summary": {
                    "type": "string"
                  }
                },
                "required": [
                  "sightings",
                  "companySignals",
                  "summary",
                  "confidence",
                  "status"
                ],
                "type": "object"
              },
              {
                "properties": {
                  "reason": {
                    "type": "string"
                  },
                  "status": {
                    "enum": [
                      "failed"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "status",
                  "reason"
                ],
                "type": "object"
              }
            ]
          }
        },
        "required": [
          "tier",
          "tone",
          "summary"
        ],
        "type": "object"
      },
      "LegitimacyTier": {
        "enum": [
          "verified",
          "clear",
          "suspicious",
          "ghost",
          "scam"
        ],
        "type": "string"
      },
...
      "Persona": {
        "enum": [
          "remote",
          "local",
          "pasted"
        ],
        "type": "string"
      },
      "Profile": {
        "properties": {
          "baseCountry": {
            "maxLength": 2,
            "minLength": 2,
            "type": "string"
          },
          "employmentPref": {
            "enum": [
              "any",
              "employee",
              "local-entity"
            ],
            "type": "string"
          },
          "relocation": {
            "$ref": "#/components/schemas/RelocationPref"
          },
          "scheduleFlex": {
            "enum": [
              "base-hours",
              "flex-evenings",
              "any-hours"
            ],
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "baseCountry",
          "relocation",
          "scheduleFlex",
          "employmentPref",
          "updatedAt"
        ],
        "type": "object"
      },
...
      "RelocationPref": {
        "enum": [
          "stay",
          "open"
        ],
        "type": "string"
      },
...
      "SourceRef": {
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "enum": [
              "ats",
              "board",
              "manual"
            ],
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "persona": {
            "$ref": "#/components/schemas/Persona"
          }
        },
        "required": [
          "id",
          "name",
          "kind",
          "persona"
        ],
        "type": "object"
      },
...
      "SummaryStripStats": {
        "properties": {
          "excluded": {
            "type": "integer"
          },
          "flagged": {
            "type": "integer"
          },
          "ghosts": {
            "type": "integer"
          },
          "scanned": {
            "type": "integer"
          },
          "sinceLast": {
            "type": "integer"
          },
          "worth": {
            "type": "integer"
          }
        },
        "required": [
          "scanned",
          "worth",
          "ghosts",
          "flagged",
          "sinceLast",
          "excluded"
        ],
        "type": "object"
      },
...
      "Tone": {
        "enum": [
          "verified",
          "good",
          "warn",
          "ghost",
          "danger",
          "neutral"
        ],
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "Auth: email+password sessions (httpOnly cookie). requireUser()/requireAdmin() guard route handlers; no Next middleware.",
    "title": "Caliber API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/applications": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "stage",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    0
                  ],
                  "type": "number"
                },
                {
                  "enum": [
                    1
                  ],
                  "type": "number"
                },
                {
                  "enum": [
                    2
                  ],
                  "type": "number"
                },
                {
                  "enum": [
                    3
                  ],
                  "type": "number"
                }
              ]
            }
          },
          {
            "in": "query",
            "name": "statusTone",
            "required": false,
            "schema": {
              "enum": [
                "good",
                "verified",
                "neutral"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "items": {
                      "items": {
                        "$ref": "#/components/schemas/Application"
                      },
                      "type": "array"
                    },
                    "nextCursor": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "items",
                    "nextCursor"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Page of tracker rows"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No session"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unknown query parameter or invalid value"
          }
        },
        "summary": "List tracker rows — filterable, cursored — F5 (B9)"
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "answersId": {
                    "type": "string"
                  },
                  "jobId": {
                    "type": "string"
                  },
                  "note": {
                    "type": "string"
                  },
                  "tailoredResumeId": {
                    "type": "string"
                  }
                },
                "required": [
                  "jobId"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Application"
                }
              }
            },
            "description": "Server sets appliedAt, stage:0, statusLabel/statusTone via features/applied/status-map.ts"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No session"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unknown job id (incl. a foreign-owned one — no existence leak)"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Duplicate jobId (details: { existingId }), job exists but isn't scored yet, or no active résumé to record"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid body"
          }
        },
        "summary": "Mark a job applied — persists a stage-0 tracker row — F5 (B9)"
      }
    },
...
    "/api/jobs": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "persona",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Persona"
            }
          },
          {
            "description": "repeatable",
            "in": "query",
            "name": "tier",
            "required": false,
            "schema": {
              "description": "repeatable",
              "items": {
                "$ref": "#/components/schemas/LegitimacyTier"
              },
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "minScore",
            "required": false,
            "schema": {
              "maximum": 5,
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "isNew",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "default 25",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "description": "default 25",
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "items": {
                      "items": {
                        "$ref": "#/components/schemas/Job"
                      },
                      "type": "array"
                    },
                    "nextCursor": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "stats": {
                      "$ref": "#/components/schemas/SummaryStripStats"
                    }
                  },
                  "required": [
                    "items",
                    "nextCursor",
                    "stats"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Page of scored jobs + hero stats computed server-side over the full scoped set"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No session"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unknown query parameter or invalid value"
          }
        },
        "summary": "Scored feed — filterable, cursored — F2 (B6)"
      }
    },
...
    "/api/profile": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            },
            "description": "The caller's profile"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No session"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The caller has no profile row yet (PUT to onboard)"
          }
        },
        "summary": "The caller's profile — base country + relocation preference"
      },
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "baseCountry": {
                    "maxLength": 2,
                    "minLength": 2,
                    "type": "string"
                  },
                  "employmentPref": {
                    "enum": [
                      "any",
                      "employee",
                      "local-entity"
                    ],
                    "type": "string"
                  },
                  "relocation": {
                    "$ref": "#/components/schemas/RelocationPref"
                  },
                  "scheduleFlex": {
                    "enum": [
                      "base-hours",
                      "flex-evenings",
                      "any-hours"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "baseCountry",
                  "relocation",
                  "scheduleFlex",
                  "employmentPref"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            },
            "description": "The created/updated profile"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No session"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid body"
          }
        },
        "summary": "Create-or-replace the caller's profile (onboarding path)"
      }
    },
...
  },
...
}